packages/apps: Update Launcher2 with performance fixes
Bind applications icons first rather than in the end when
all application view is visible. This gives performance improvement
in orientation change. Also, crash is observed when boot up with keypad
slided. This crash is caused due to activity getting orientation change
event before it creates allocation. Text is also seen as shifted in this
mode, change the dimensions to make it fix the same.
This addresses the CR's 230338

Change-Id:
CRs-fixed: 230335, 230336